refactor: cleanup unparser references
