+++ packages/runtime/src/binding/expression-parser.ts
@@ -1,46 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
 import {
-  AccessKeyedExpression,
-  AccessMemberExpression,
-  AccessScopeExpression,
-  AccessThisExpression,
-  ArrayBindingPattern,
-  ArrayLiteralExpression,
-  AssignExpression,
-  BinaryExpression,
-  BindingBehaviorExpression,
-  BindingIdentifier,
-  CallFunctionExpression,
-  CallMemberExpression,
-  CallScopeExpression,
-  ConditionalExpression,
-  CustomExpression,
-  ForOfStatement,
-  Interpolation,
-  ObjectBindingPattern,
-  ObjectLiteralExpression,
-  PrimitiveLiteralExpression,
-  TaggedTemplateExpression,
-  TemplateExpression,
-  UnaryExpression,
-  ValueConverterExpression,
-  AnyBindingExpression,
-  BinaryOperator,
-  BindingIdentifierOrPattern,
-  IsAssign,
-  IsAssignable,
-  IsBinary,
-  IsBindingBehavior,
-  IsExpressionOrStatement,
-  IsLeftHandSide,
-  IsValueConverter,
-  UnaryOperator,
-  DestructuringAssignmentSingleExpression as DASE,
-  DestructuringAssignmentExpression as DAE,
-  ArrowFunction,
-  AccessGlobalExpression,
-  CallGlobalExpression,
-  type ExpressionKind,
   ekAccessThis,
   ekAccessGlobal,
   ekAccessMember,
@@ -50,10 +9,53 @@ import {
   ekObjectBindingPattern,
   ekBindingIdentifier,
   ekObjectDestructuring,
-  AccessBoundaryExpression,
-} from './ast';
+} from './expression-types';
 import { createInterface, createLookup, objectAssign } from '../utilities';
 import { ErrorNames, createMappedError } from '../errors';
+import {
+  AccessBoundaryExpression,
+  AccessGlobalExpression,
+  AccessKeyedExpression,
+  AccessMemberExpression,
+  AccessScopeExpression,
+  AccessThisExpression,
+  AnyBindingExpression,
+  ArrayBindingPattern,
+  ArrayLiteralExpression,
+  ArrowFunction,
+  AssignExpression,
+  BinaryExpression,
+  BinaryOperator,
+  BindingBehaviorExpression,
+  BindingIdentifier,
+  BindingIdentifierOrPattern,
+  CallFunctionExpression,
+  CallGlobalExpression,
+  CallMemberExpression,
+  CallScopeExpression,
+  ConditionalExpression,
+  CustomExpression,
+  DestructuringAssignmentExpression as DAE,
+  DestructuringAssignmentSingleExpression as DASE,
+  ExpressionKind,
+  ForOfStatement,
+  Interpolation,
+  IsAssign,
+  IsAssignable,
+  IsBinary,
+  IsBindingBehavior,
+  IsExpressionOrStatement,
+  IsLeftHandSide,
+  IsValueConverter,
+  ObjectBindingPattern,
+  ObjectLiteralExpression,
+  PrimitiveLiteralExpression,
+  TaggedTemplateExpression,
+  TemplateExpression,
+  UnaryExpression,
+  UnaryOperator,
+  ValueConverterExpression,
+} from '@aurelia/expression-parser';
 
 export interface IExpressionParser extends ExpressionParser {}
 export const IExpressionParser = createInterface<IExpressionParser>('IExpressionParser', x => x.singleton(ExpressionParser));
