+++ packages/runtime/src/binding/expression-types.ts
@@ -1,0 +1,31 @@
+/** @internal */ export const ekAccessThis = 'AccessThis';
+/** @internal */ export const ekAccessBoundary = 'AccessBoundary';
+/** @internal */ export const ekAccessGlobal = 'AccessGlobal';
+/** @internal */ export const ekAccessScope = 'AccessScope';
+/** @internal */ export const ekArrayLiteral = 'ArrayLiteral';
+/** @internal */ export const ekObjectLiteral = 'ObjectLiteral';
+/** @internal */ export const ekPrimitiveLiteral = 'PrimitiveLiteral';
+/** @internal */ export const ekTemplate = 'Template';
+/** @internal */ export const ekUnary = 'Unary';
+/** @internal */ export const ekCallScope = 'CallScope';
+/** @internal */ export const ekCallMember = 'CallMember';
+/** @internal */ export const ekCallFunction = 'CallFunction';
+/** @internal */ export const ekCallGlobal = 'CallGlobal';
+/** @internal */ export const ekAccessMember = 'AccessMember';
+/** @internal */ export const ekAccessKeyed = 'AccessKeyed';
+/** @internal */ export const ekTaggedTemplate = 'TaggedTemplate';
+/** @internal */ export const ekBinary = 'Binary';
+/** @internal */ export const ekConditional = 'Conditional';
+/** @internal */ export const ekAssign = 'Assign';
+/** @internal */ export const ekArrowFunction = 'ArrowFunction';
+/** @internal */ export const ekValueConverter = 'ValueConverter';
+/** @internal */ export const ekBindingBehavior = 'BindingBehavior';
+/** @internal */ export const ekArrayBindingPattern = 'ArrayBindingPattern';
+/** @internal */ export const ekObjectBindingPattern = 'ObjectBindingPattern';
+/** @internal */ export const ekBindingIdentifier = 'BindingIdentifier';
+/** @internal */ export const ekForOfStatement = 'ForOfStatement';
+/** @internal */ export const ekInterpolation = 'Interpolation';
+/** @internal */ export const ekArrayDestructuring = 'ArrayDestructuring';
+/** @internal */ export const ekObjectDestructuring = 'ObjectDestructuring';
+/** @internal */ export const ekDestructuringAssignmentLeaf = 'DestructuringAssignmentLeaf';
+/** @internal */ export const ekCustom = 'Custom';
