+++ packages/runtime/src/index.ts
@@ -1,62 +1,62 @@
 export {
-  type ExpressionKind,
-  // Ast nodes
-  CallFunctionExpression,
-  CustomExpression,
-  BindingBehaviorExpression,
-  ValueConverterExpression,
-  AssignExpression,
-  ConditionalExpression,
-  AccessThisExpression,
-  AccessGlobalExpression,
-  AccessScopeExpression,
-  AccessBoundaryExpression,
-  AccessMemberExpression,
-  AccessKeyedExpression,
-  CallScopeExpression,
-  CallMemberExpression,
-  BinaryExpression,
-  UnaryExpression,
-  PrimitiveLiteralExpression,
-  ArrayLiteralExpression,
-  ObjectLiteralExpression,
-  TemplateExpression,
-  TaggedTemplateExpression,
-  ArrayBindingPattern,
-  ObjectBindingPattern,
-  BindingIdentifier,
-  ForOfStatement,
-  Interpolation,
-  DestructuringAssignmentExpression,
-  DestructuringAssignmentSingleExpression,
-  DestructuringAssignmentRestExpression,
-  ArrowFunction,
+  // type ExpressionKind,
+  // // Ast nodes
+  // CallFunctionExpression,
+  // CustomExpression,
+  // BindingBehaviorExpression,
+  // ValueConverterExpression,
+  // AssignExpression,
+  // ConditionalExpression,
+  // AccessThisExpression,
+  // AccessGlobalExpression,
+  // AccessScopeExpression,
+  // AccessBoundaryExpression,
+  // AccessMemberExpression,
+  // AccessKeyedExpression,
+  // CallScopeExpression,
+  // CallMemberExpression,
+  // BinaryExpression,
+  // UnaryExpression,
+  // PrimitiveLiteralExpression,
+  // ArrayLiteralExpression,
+  // ObjectLiteralExpression,
+  // TemplateExpression,
+  // TaggedTemplateExpression,
+  // ArrayBindingPattern,
+  // ObjectBindingPattern,
+  // BindingIdentifier,
+  // ForOfStatement,
+  // Interpolation,
+  // DestructuringAssignmentExpression,
+  // DestructuringAssignmentSingleExpression,
+  // DestructuringAssignmentRestExpression,
+  // ArrowFunction,
 
-  astVisit,
-  Unparser,
+  // astVisit,
+  // Unparser,
 
   // ast typing helpers
-  type AnyBindingExpression,
+  // type AnyBindingExpression,
   type BindingBehaviorInstance,
-  type IsPrimary,
-  type IsLiteral,
-  type IsLeftHandSide,
-  type IsUnary,
-  type IsBinary,
-  type IsConditional,
-  type IsAssign,
-  type IsValueConverter,
-  type IsBindingBehavior,
-  type IsAssignable,
-  type IsExpression,
-  type IsExpressionOrStatement,
-  type IVisitor,
-  type BinaryOperator,
-  type BindingIdentifierOrPattern,
-  type UnaryOperator,
+  // type IsPrimary,
+  // type IsLiteral,
+  // type IsLeftHandSide,
+  // type IsUnary,
+  // type IsBinary,
+  // type IsConditional,
+  // type IsAssign,
+  // type IsValueConverter,
+  // type IsBindingBehavior,
+  // type IsAssignable,
+  // type IsExpression,
+  // type IsExpressionOrStatement,
+  // type IVisitor,
+  // type BinaryOperator,
+  // type BindingIdentifierOrPattern,
+  // type UnaryOperator,
   type IAstEvaluator,
   type ValueConverterInstance,
-} from './binding/ast';
+} from './binding/ast.eval';
 export {
   astAssign,
   astBind,
